add task instance core module
