+++ kubetask/core/task.py
@@ -1,7 +1,13 @@
-from .task_instance import TaskInstance
+# from .task_instance import TaskInstance
 from .constants import State, Priority
-from kubetask.models.model import TaskModel
+from kubetask.models.model import TaskModel, TaskInstanceModel
 from kubetask.core.db import DB
+
+def update_state(func) : 
+    def caller(self) : 
+        func(self) 
+        self.model_obj.state = self.state
+    return caller 
 
 class Task:
     def __init__(self, task_name, docker_url, command, schedule_at=None, start_at=None, priority=None, task_id=None):
@@ -23,12 +29,6 @@ class Task:
 
         self.task_id = self.model_obj.task_id
 
-    def update_state(func) : 
-        def caller(self) : 
-            func(self) 
-            self.model_obj.state = self.state
-        return caller 
-
     @update_state
     def schedule(self):
         self.state = State.SCHEDULED
@@ -42,3 +42,42 @@ class Task:
         self.state = State.STARTED
         task_instance = TaskInstance(self)
         return task_instance
+
+
+
+class TaskInstance:
+    def __init__(self, task, id=None):
+        self.id = id
+        self.task = task
+        self.state = State.NOT_STARTED
+        self.start_ts = None
+        self.end_ts = None
+        self.model_obj = None
+
+    def create_db_object(self):
+        kwargs = vars(self).copy()
+        kwargs["task"] = self.task.model_obj
+        kwargs.pop("model_obj")
+        self.model_obj = DB.create(TaskInstanceModel, kwargs)
+
+    @update_state
+    def start(self):
+        """starts the execution of the task
+        1. push the task to the queue
+        2. Update DB with task instance details
+        """
+        self.state = State.STARTED
+        self.create_db_object()       
+        
+
+    @update_state
+    def complete(self):
+        """Update DB that the task is complete
+        """
+        if not self.model_obj:
+            raise Exception("Task instance not started yet")
+        self.state = State.COMPLETED
+
+    
+
+
